requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
allure-pytest==2.8.18
allure-python-commons==2.8.18
amqp==2.6.1
anyjson==0.3.3
appscript==1.1.2
APScheduler==3.6.3
asgiref==3.2.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.4.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.15.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.1
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
constantly==15.1.0
contextlib2==21.6.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
crypto==1.4.1
cryptography==36.0.2
daphne==2.5.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==2.2.13
django-apscheduler==0.5.1
django-celery==3.3.1
django-celery-results==2.2.0
django-cors-headers==3.4.0
django-easy-audit==1.3.0
django-filter==2.3.0
django-multiselectfield==0.1.12
django-redis==4.12.1
django-relativedelta==1.1.2
django-simpleui==2020.9
django-user-agents==0.4.0
djangorestframework==3.11.0
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
drf-writable-nested==0.6.1
drf-yasg==1.17.1
enum34==1.1.10
esdk-obs-python==3.21.4
et-xmlfile==1.1.0
eventlet==0.26.1
ffmpy==0.2.3
gevent==21.12.0
greenlet==1.1.2
hiredis==1.1.0
hyperlink==19.0.0
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
inflection==0.5.0
iniconfig==1.0.1
itypes==1.2.0
Jinja2==2.11.2
jira==3.0.1
jmespath==0.10.0
keyring==23.2.1
kombu==4.6.11
lxml==4.5.2
MarkupSafe==1.1.1
monotonic==1.5
more-itertools==8.5.0
msgpack==0.6.2
Naked==0.1.31
namedlist==1.8
oauthlib==3.1.1
openpyxl==3.0.9
oss2==2.15.0
packaging==20.4
Pillow==9.0.1
pluggy==0.13.1
prettytable==2.1.0
prompt-toolkit==3.0.20
psutil==5.8.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.10.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-ordering==0.6
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytils==0.3
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
schema==0.7.4
selenium==3.141.0
service-identity==18.1.0
shellescape==3.8.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
tomli==2.0.1
tools==0.1.9
Twisted==20.3.0
txaio==20.4.1
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==2.1
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.11
user-agents==2.2.0
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
